Add Structure component and Filter component
