+++ client/src/components/Structure/Structure.css
@@ -1,0 +1,34 @@
+.structure-container {
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	justify-content: center;
+	width: 100%;
+	height: calc(100% - 61px);
+	gap: 10px;
+}
+
+.structure-left,
+.structure-right {
+	display: flex;
+	flex-direction: column;
+	gap: 10px;
+	border-radius: 1rem;
+	background: var(--color-bg-variant);
+	box-shadow: 0 -5px 10px rgba(0, 0, 0, 0.5);
+	height: 100%;
+	overflow-y: auto;
+	align-items: center;
+	padding: 10px;
+	border: 1px solid var(--color-bg);
+}
+
+.structure-left {
+	width: 20%;
+	gap: 20px;
+}
+
+.structure-right {
+	width: 100%;
+	padding: 10px;
+}
