+++ client/src/components/Structure/Structure.jsx
@@ -1,0 +1,13 @@
+import './Structure.css';
+
+const Structure = ({ LeftCompnonet, RightComponent, ContainerComponent }) => {
+	return (
+		<div className="structure-container">
+			{ContainerComponent}
+			<div className="structure-left">{LeftCompnonet}</div>
+			<div className="structure-right">{RightComponent}</div>
+		</div>
+	);
+};
+
+export default Structure;
