+++ client/src/components/Student/StudentTable.css
@@ -1,38 +1,3 @@
-.students-container {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	justify-content: center;
-	width: 100%;
-	height: calc(100% - 61px);
-	gap: 10px;
-}
-
-.students-left,
-.students-right {
-	display: flex;
-	flex-direction: column;
-	gap: 10px;
-	border-radius: 1rem;
-	background: var(--color-bg-variant);
-	box-shadow: 0 -5px 10px rgba(0, 0, 0, 0.5);
-	height: 100%;
-	overflow-y: auto;
-	align-items: center;
-	padding: 10px;
-	border: 1px solid var(--color-bg);
-}
-
-.students-left {
-	width: 20%;
-	gap: 20px;
-}
-
-.students-right {
-	width: 100%;
-	padding: 10px;
-}
-
 .verify-button {
 	background-color: transparent;
 	display: flex;
